pydantic-llm-mixin/pydantic-llm-mixin/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pydantic-llm-mixin"
version = "0.1.0"
description = "Best-of-breed GenerativeMixin for Pydantic models - turn any model into an LLM-generated instance with robust validation"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Alan Eyzaguirre", email = "alan@eyzaguirre.co" }
]
keywords = ["pydantic", "llm", "groq", "generative-ai", "validation", "json-schema"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic",
    "httpx",
    "rich",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
    "typer>=0.20.0",
    "geopy>=2.4.1",
    "email-validator>=2.3.0",
    "itsdangerous>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "ty",
]
railway = [
    "fastapi",
    "uvicorn",
]

[project.urls]
Homepage = "https://github.com/eyzaguirre-co/pydantic-llm-mixin"
Repository = "https://github.com/eyzaguirre-co/pydantic-llm-mixin"
Issues = "https://github.com/eyzaguirre-co/pydantic-llm-mixin/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/pydantic_llm_mixin"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

